+++ src/public.ts
@@ -114,8 +114,12 @@ export {
 } from './actions/public/watchBlocks.js'
 export {
   watchEvent,
-  type OnLogsFn,
-  type OnLogsParameter,
+  type WatchEventOnLogsFn,
+  /** @deprecated - use `WatchEventOnLogsFn` instead. */
+  type WatchEventOnLogsFn as OnLogsFn,
+  type WatchEventOnLogsParameter,
+  /** @deprecated - use `WatchEventOnLogsParameter` instead. */
+  type WatchEventOnLogsParameter as OnLogsParameter,
 } from './actions/public/watchEvent.js'
 export {
   watchPendingTransactions,