+++ src/types/filter.ts
@@ -1,5 +1,6 @@
 import type { Abi } from 'abitype'
 
+import type { BlockNumber, BlockTag } from './block.js'
 import type { MaybeExtractEventArgsFromAbi } from './contract.js'
 import type { EIP1193RequestFn, PublicRpcSchema } from './eip1193.js'
 import type { Hex } from './misc.js'
@@ -16,42 +17,47 @@ type FilterRpcSchema = Filter_<
 
 export type Filter<
   TFilterType extends FilterType = 'event',
-  TAbi extends Abi | readonly unknown[] = Abi,
+  TAbi extends Abi | readonly unknown[] | undefined = undefined,
   TEventName extends string | undefined = undefined,
   TArgs extends
     | MaybeExtractEventArgsFromAbi<TAbi, TEventName>
     | undefined = MaybeExtractEventArgsFromAbi<TAbi, TEventName>,
   TStrict extends boolean | undefined = undefined,
+  TFromBlock extends BlockNumber | BlockTag | undefined = undefined,
+  TToBlock extends BlockNumber | BlockTag | undefined = undefined,
 > = {
   id: Hex
   request: EIP1193RequestFn<FilterRpcSchema>
   type: TFilterType
 } & (TFilterType extends 'event'
-  ? TAbi extends Abi
-    ? undefined extends TEventName
-      ? {
-          abi: TAbi
+  ? {
+      fromBlock?: TFromBlock
+      toBlock?: TToBlock
+    } & (TAbi extends Abi
+      ? undefined extends TEventName
+        ? {
+            abi: TAbi
+            args?: never
+            eventName?: never
+            strict: TStrict
+          }
+        : TArgs extends MaybeExtractEventArgsFromAbi<TAbi, TEventName>
+        ? {
+            abi: TAbi
+            args: TArgs
+            eventName: TEventName
+            strict: TStrict
+          }
+        : {
+            abi: TAbi
+            args?: never
+            eventName: TEventName
+            strict: TStrict
+          }
+      : {
+          abi?: never
           args?: never
           eventName?: never
-          strict: TStrict
-        }
-      : TArgs extends MaybeExtractEventArgsFromAbi<TAbi, TEventName>
-      ? {
-          abi: TAbi
-          args: TArgs
-          eventName: TEventName
-          strict: TStrict
-        }
-      : {
-          abi: TAbi
-          args?: never
-          eventName: TEventName
-          strict: TStrict
-        }
-    : {
-        abi?: never
-        args?: never
-        eventName?: never
-        strict?: never
-      }
+          strict?: never
+        })
   : {})